Move entities builders to separate package.
